requirements.txt
torch
torchvision
numpy
pillow
tqdm
tb-nightly
future
natsort
opencv-python
